requirements.txt
mysql-connector-python==8.0.22
pygame==2.0.1
